refactor(reactor): move secret redaction to a keyless module
redactSecrets and redactError lived inside the agent-render provider
adapter, which statically imports @openai/agents, so nothing on the
keyless side (the reconciler core, the offline CLI surfaces) could
scrub text without crossing the model-bearing boundary. The scrubbers
are pure string logic with no imports, so they now live in a top-level
keyless module; the adapter re-exports them for its existing call
sites. No behavior change.
